tab5: the Tab5 firmware — hardware layer, ESP-IDF project, docs and shots
hal_tab5.c on Espressif's Tab5 BSP: panel detection for all three revisions,
PPA-rotated landscape frames flipped between two DPI buffers on vsync, touch,
BMI270, INA226, RX8130, ES8388/ES7210 audio workers, the SC2356 camera through
esp_video with PPA scaling and JPEG snapshots, microSD, Wi-Fi through the C6 over
esp-hosted, and a listen-only TWAI CAN tap on Grove. Builds against ESP-IDF 5.5.1
with the component versions pinned in the manifests; not yet run on a device.

Also: hue-aware RGB565 token rounding, sparkline minimum spans, the level and
control-center fixes the simulator tour found, a README with the screens, and a
short section in AGENTS.md.

+++ tab5/components/catalyst/src/cat_logs.c
@@ -255,7 +255,7 @@ static bool dslog(const uint8_t *d, size_t len, cat_log_t *o)
     }
     uint32_t ver = (uint32_t)be(d, 4);
     if (ver != 3 && ver != 4) {
-        snprintf(o->error, sizeof o->error, "dslog version %u", ver);
+        snprintf(o->error, sizeof o->error, "dslog version %u", (unsigned)ver);
         return false;
     }
     o->start_unix = (int64_t)be(d + 4, 8) - LABVIEW_EPOCH;
